refactor: extract EventCoverImageField from CreateEditEventFields
First template-decomposition increment for the event form. Move the entire
cover-image concern — the 3-state upload UI plus its signed-URL upload logic —
into a self-contained child component wired by a single prop/event:

  <EventCoverImageField :image-url="..." @update="...coverImageURL..." />

Parent drops ~190 lines (1045 -> 857) and several now-unused imports
(AddImage, useMutation/CREATE_SIGNED_STORAGE_URL, upload utils, useUsername).
Existing CreateEditEventFields tests still pass; adds focused tests for the
new child.

Also fixes a latent bug carried over from the original: the loading spinner
was left stuck on when an upload returned no link; the child now resets it.
